All libraries moved to libft.h file. ft_itoabase implemented.

+++ ft_strndup.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_strndup.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: msales-a <msales-a@student.42.fr>          +#+  +:+       +#+        */
+/*   By: msales-a <msales-a@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/01/25 13:08:21 by msales-a          #+#    #+#             */
-/*   Updated: 2020/02/08 18:00:08 by msales-a         ###   ########.fr       */
+/*   Updated: 2020/04/19 17:43:38 by msales-a         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,11 +14,11 @@
 
 char	*ft_strndup(char const *s, size_t n)
 {
-	int		index;
+	size_t	index;
 	char	*dst;
 
 	index = ft_strlen(s);
-	index = ((int)n < index) ? n : index;
+	index = (n < index) ? n : index;
 	if (!(dst = (char*)ft_calloc((index + 1), sizeof(char))))
 		return (0);
 	ft_strlcpy(dst, s, index + 1);
