+++ ft_strnlen.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_strnlen.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: msales-a <marvin@42.fr>                    +#+  +:+       +#+        */
+/*   By: msales-a <msales-a@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/01/25 16:38:17 by msales-a          #+#    #+#             */
-/*   Updated: 2020/01/25 16:59:43 by msales-a         ###   ########.fr       */
+/*   Updated: 2020/04/19 17:42:21 by msales-a         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 size_t	ft_strnlen(const char *s, size_t maxlen)
 {
-	int len;
+	size_t len;
 
 	len = ft_strlen(s);
-	return ((len < (int)maxlen) ? len : maxlen);
+	return ((len < maxlen) ? len : maxlen);
 }
